// src/app/models/comentarios.ts
export class Comentarios {
    $key: string;
    id_noticia: string;
    username: string;
    date: string;
    comment: string;
    userPicture: string;
    uid: string;
}

export class comentariosConNoticia {
    $key: string;
    comment: string;
    date: string;
    userPicture: string;
    username: string;
    uid: string;
}